Moved `UIViewController` activation logic out of the `Router` to `StackPresentationHandler`
